Example showing best practice of keeping API KEY private
Some of us are less proficient in ruby, and it took me a while
to figure out how to read my environment variable from the vagrant
file. I'm a bit wiser for this, and hopefuly it will save someone
a bit of time.

Hopefuly, also, it will prevent a couple of publicly spilled
api keys...

Changes to be committed:
        modified:   README.md